pyproject.toml
[tool.poetry]
name = "workshop2"
version = "0.1.0"
description = "struct a Data Analytics project from Jupyter to Deployment"
authors = ["PedroPauloMR <pedrop.ribeiro@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
taskipy = "^1.12.2"
pytest = "^8.0.0"
httpx = "^0.26.0"
pydantic = "^2.5.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "uvicorn app.main:app --reload"
test = "pytest -v"
kill = "fuser -k 8000/tcp"
